Restructed solution
