+++ Octopus.Conductor.sln
@@ -1,0 +1,74 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31919.166
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{50A7155F-1A28-4ED6-9DC8-00637D741065}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Octopus.Conductor.Domain", "src\Core\Octopus.Conductor.Domain\Octopus.Conductor.Domain.csproj", "{ABCD5411-4792-40AB-95C0-14CBB770C81C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Octopus.Conductor.Infrastructure.RelationalDB", "src\Infrastructure\Octopus.Conductor.Infrastructure.RelationalDB\Octopus.Conductor.Infrastructure.RelationalDB.csproj", "{711CE993-C06C-446E-A85D-CCCE7F3A249A}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Octopus.Conductor.WebApi", "src\Presentation\Octopus.Conductor.WebApi\Octopus.Conductor.WebApi.csproj", "{725E6E64-75FC-494E-897D-F3EE3134CA11}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{77C393FB-BEE3-4E46-8B10-3555BC545675}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Infrastructure", "Infrastructure", "{BD4E626C-1B38-49D8-8217-394DA4158EB4}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Presentation", "Presentation", "{5ED6A24B-E342-4AEE-BE49-265232C10326}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Octopus.Conductor.Application", "src\Core\Octopus.Conductor.Application\Octopus.Conductor.Application.csproj", "{0F38CB86-B8E2-4B2F-9A40-04CF02A40CC8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Octopus.Conductor.Infrastructure.Services", "src\Infrastructure\Octopus.Conductor.Infrastructure.Services\Octopus.Conductor.Infrastructure.Services.csproj", "{B9EADB5B-8AF1-429F-9E09-42DF435BC58E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Octopus.Conductor.Infrastructure.WorkerService", "src\Infrastructure\Octopus.Conductor.Infrastructure.WorkerService\Octopus.Conductor.Infrastructure.WorkerService.csproj", "{11924E27-D08A-4503-A05F-2B929383C818}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{ABCD5411-4792-40AB-95C0-14CBB770C81C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ABCD5411-4792-40AB-95C0-14CBB770C81C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ABCD5411-4792-40AB-95C0-14CBB770C81C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ABCD5411-4792-40AB-95C0-14CBB770C81C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{711CE993-C06C-446E-A85D-CCCE7F3A249A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{711CE993-C06C-446E-A85D-CCCE7F3A249A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{711CE993-C06C-446E-A85D-CCCE7F3A249A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{711CE993-C06C-446E-A85D-CCCE7F3A249A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{725E6E64-75FC-494E-897D-F3EE3134CA11}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{725E6E64-75FC-494E-897D-F3EE3134CA11}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{725E6E64-75FC-494E-897D-F3EE3134CA11}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{725E6E64-75FC-494E-897D-F3EE3134CA11}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0F38CB86-B8E2-4B2F-9A40-04CF02A40CC8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0F38CB86-B8E2-4B2F-9A40-04CF02A40CC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0F38CB86-B8E2-4B2F-9A40-04CF02A40CC8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0F38CB86-B8E2-4B2F-9A40-04CF02A40CC8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B9EADB5B-8AF1-429F-9E09-42DF435BC58E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B9EADB5B-8AF1-429F-9E09-42DF435BC58E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B9EADB5B-8AF1-429F-9E09-42DF435BC58E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B9EADB5B-8AF1-429F-9E09-42DF435BC58E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{11924E27-D08A-4503-A05F-2B929383C818}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{11924E27-D08A-4503-A05F-2B929383C818}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{11924E27-D08A-4503-A05F-2B929383C818}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{11924E27-D08A-4503-A05F-2B929383C818}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{ABCD5411-4792-40AB-95C0-14CBB770C81C} = {77C393FB-BEE3-4E46-8B10-3555BC545675}
+		{711CE993-C06C-446E-A85D-CCCE7F3A249A} = {BD4E626C-1B38-49D8-8217-394DA4158EB4}
+		{725E6E64-75FC-494E-897D-F3EE3134CA11} = {5ED6A24B-E342-4AEE-BE49-265232C10326}
+		{77C393FB-BEE3-4E46-8B10-3555BC545675} = {50A7155F-1A28-4ED6-9DC8-00637D741065}
+		{BD4E626C-1B38-49D8-8217-394DA4158EB4} = {50A7155F-1A28-4ED6-9DC8-00637D741065}
+		{5ED6A24B-E342-4AEE-BE49-265232C10326} = {50A7155F-1A28-4ED6-9DC8-00637D741065}
+		{0F38CB86-B8E2-4B2F-9A40-04CF02A40CC8} = {77C393FB-BEE3-4E46-8B10-3555BC545675}
+		{B9EADB5B-8AF1-429F-9E09-42DF435BC58E} = {BD4E626C-1B38-49D8-8217-394DA4158EB4}
+		{11924E27-D08A-4503-A05F-2B929383C818} = {BD4E626C-1B38-49D8-8217-394DA4158EB4}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B5037395-1809-4148-A8EB-0523B0AAC1FE}
+	EndGlobalSection
+EndGlobal
