+++ src/Core/Octopus.Conductor.Application/Interfaces/IFolderListener.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace Octopus.Conductor.Application.Interfaces
+{
+    public interface IFolderListener
+    {
+        public Task MoveEntityFilesAsync(CancellationToken cancellationToken = default);
+    }
+}
