+++ src/Core/Octopus.Conductor.Application/Interfaces/IRepository.cs
@@ -1,0 +1,21 @@
+﻿using Octopus.Conductor.Domain.Common;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Octopus.Conductor.Application.Interfaces
+{
+    public interface IRepository
+    {
+        Task<IEnumerable<TEntity>> GetAllAsync<TEntity>(
+            CancellationToken cancellationToken = default) where TEntity : BaseEntity;
+        Task<TEntity> AddAsync<TEntity>(
+            TEntity entity,
+            CancellationToken cancellationToken = default) where TEntity : BaseEntity;
+        Task<TEntity> GetByIdAsync<TEntity>(
+            int id,
+            CancellationToken cancellationToken = default) where TEntity : BaseEntity;
+        void Delete<TEntity>(TEntity entity) where TEntity : BaseEntity;
+        Task<int> SaveChangesAsync(CancellationToken cancellationToken = default);
+    }
+}
