+++ src/Core/Octopus.Conductor.Application/Octopus.Conductor.Application.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Octopus.Conductor.Domain\Octopus.Conductor.Domain.csproj" />
+  </ItemGroup>
+
+</Project>
