+++ src/Core/Octopus.Conductor.Domain/Common/BaseEntity.cs
@@ -1,0 +1,7 @@
+﻿namespace Octopus.Conductor.Domain.Common
+{
+    public abstract class BaseEntity
+    {
+        public int Id { get; set; }
+    }
+}
