+++ src/Core/Octopus.Conductor.Domain/Entities/ConductorEntityDescription.cs
@@ -1,0 +1,11 @@
+﻿using Octopus.Conductor.Domain.Common;
+
+namespace Octopus.Conductor.Domain.Entities
+{
+    public class ConductorEntityDescription : BaseEntity
+    {
+        public string EntityType { get; set; }
+        public string InputDirectory { get; set; }
+        public string OutputDirectory { get; set; }
+    }
+}
