+++ src/Core/Octopus.Conductor.Domain/Octopus.Conductor.Domain.csproj
