+++ src/Infrastructure/Octopus.Conductor.Infrastructure.Services/Octopus.Conductor.Infrastructure.Services.csproj
@@ -1,0 +1,7 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+</Project>
