+++ src/Infrastructure/Octopus.Conductor.Infrastructure.WorkerService/Interfaces/IExtendedHostingService.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.Extensions.Hosting;
+using System.Threading.Tasks;
+
+namespace Octopus.Conductor.Infrastructure.WorkerService.Interfaces
+{
+    public interface IExtendedHostingService : IHostedService
+    {
+        public TaskStatus GetWorkerStatus();
+    }
+}
