--- src/Octopus.Conductor.Core/Common/BaseEntity.cs
@@ -1,7 +1,0 @@
-﻿namespace Octopus.Conductor.Core.Common
-{
-    public abstract class BaseEntity
-    {
-        public int Id { get; set; }
-    }
-}
