--- src/Octopus.Conductor.Core/Entities/ConductorEntityDescription.cs
@@ -1,11 +1,0 @@
-﻿using Octopus.Conductor.Core.Common;
-
-namespace Octopus.Conductor.Core.Entities
-{
-    public class ConductorEntityDescription : BaseEntity
-    {
-        public string EntityType { get; set; }
-        public string InputDirectory { get; set; }
-        public string OutputDirectory { get; set; }
-    }
-}
