--- src/Octopus.Conductor.Infrastructure/Hosting/IExtendedHostingService.cs
@@ -1,14 +1,0 @@
-﻿using Microsoft.Extensions.Hosting;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Octopus.Conductor.Infrastructure.Hosting
-{
-    public interface IExtendedHostingService : IHostedService
-    {
-        public TaskStatus GetWorkerStatus();
-    }
-}
