+++ src/Presentation/Octopus.Conductor.WebApi/Octopus.Conductor.WebApi.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="11.0.0" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Core\Octopus.Conductor.Application\Octopus.Conductor.Application.csproj" />
+    <ProjectReference Include="..\..\Infrastructure\Octopus.Conductor.Infrastructure.RelationalDB\Octopus.Conductor.Infrastructure.RelationalDB.csproj" />
+  </ItemGroup>
+
+</Project>
