+++ src/Presentation/Octopus.Conductor.WebApi/Profiles/EntityDescriptionProfile.cs
@@ -1,0 +1,15 @@
+﻿using AutoMapper;
+using Octopus.Conductor.Domain.Entities;
+using Octopus.Conductor.WebApi.Models;
+
+namespace Octopus.Conductor.WebApi.Profiles
+{
+    public class EntityDescriptionProfile : Profile
+    {
+        public EntityDescriptionProfile()
+        {
+            CreateMap<CreateEntityDescriptionDto, ConductorEntityDescription>();
+            CreateMap<ConductorEntityDescription, ReadEntityDescriptionDto>();
+        }
+    }
+}
