+++ src/Presentation/Octopus.Conductor.WebApi/Program.cs
@@ -1,0 +1,26 @@
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Octopus.Conductor.WebApi.Services;
+
+namespace Octopus.Conductor.WebApi
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            CreateHostBuilder(args).Build().Run();
+        }
+
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                    {
+                        webBuilder.UseStartup<Startup>();
+                    })
+                .ConfigureServices(services =>
+                    {
+                        services.AddHostedService<FileMovingBackgroundService>();
+                    });
+    }
+}
