+++ src/Presentation/Octopus.Conductor.WebApi/Settings/WorkerSettings.cs
